Reorganize forms code and more cleanup
