+++ recipe-server/client/control_new/components/App.js
@@ -12,50 +12,55 @@ import QueryServiceInfo from 'control_new/components/data/QueryServiceInfo';
 
 const { Content, Header, Sider } = Layout;
 
-export default function App({ children }) {
-  return (
-    <LocaleProvider locale={enUS}>
-      <Layout>
-        {/*
-          Global query components; add any queries for data needed across the
-          entire app that we only need to fetch once.
-        */}
-        <QueryActions />
-        <QueryServiceInfo />
 
-        <Header>
-          <CurrentUserDetails />
-          <div className="logo">
-            <Link href="/">SHIELD Control Panel</Link>
-          </div>
-        </Header>
+export default class App extends React.Component {
+  static propTypes = {
+    children: PropTypes.node,
+  };
 
+  static defaultProps = {
+    children: null,
+  };
+
+  render() {
+    const { children } = this.props;
+
+    return (
+      <LocaleProvider locale={enUS}>
         <Layout>
-          <Sider
-            className="sidebar"
-            breakpoint="sm"
-            collapsedWidth="0"
-          >
-            <NavigationMenu />
-          </Sider>
+          {/*
+           Global query components; add any queries for data needed across the
+           entire app that we only need to fetch once.
+           */}
+          <QueryActions />
+          <QueryServiceInfo />
 
-          <Layout className="content-wrapper">
-            <NavigationCrumbs />
+          <Header>
+            <CurrentUserDetails />
+            <div className="logo">
+              <Link href="/">SHIELD Control Panel</Link>
+            </div>
+          </Header>
 
-            <Content className="content">
-              {children}
-            </Content>
+          <Layout>
+            <Sider
+              className="sidebar"
+              breakpoint="sm"
+              collapsedWidth="0"
+            >
+              <NavigationMenu />
+            </Sider>
+
+            <Layout className="content-wrapper">
+              <NavigationCrumbs />
+
+              <Content className="content">
+                {children}
+              </Content>
+            </Layout>
           </Layout>
         </Layout>
-      </Layout>
-    </LocaleProvider>
-  );
+      </LocaleProvider>
+    );
+  }
 }
-
-App.propTypes = {
-  children: PropTypes.node,
-};
-
-App.defaultProps = {
-  children: null,
-};
