+++ recipe-server/client/control_new/components/common/BooleanIcon.js
@@ -3,12 +3,15 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 
-export default function BooleanIcon({ value }) {
-  const type = value ? 'check' : 'close';
-  const booleanClass = value ? 'is-true' : 'is-false';
-  return <Icon className={`boolean-icon ${booleanClass}`} type={type} />;
-}
+export default class BooleanIcon extends React.Component {
+  static propTypes = {
+    value: PropTypes.bool.isRequired,
+  };
 
-BooleanIcon.propTypes = {
-  value: PropTypes.bool.isRequired,
-};
+  render() {
+    const { value } = this.props;
+    const type = value ? 'check' : 'close';
+    const booleanClass = value ? 'is-true' : 'is-false';
+    return <Icon className={`boolean-icon ${booleanClass}`} type={type}/>;
+  }
+}
