+++ recipe-server/client/control_new/components/common/CheckboxMenu.js
@@ -3,32 +3,36 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 
-export default function CheckboxMenu({ checkboxes, label, onChange, options }) {
-  const menu = (
-    <Checkbox.Group
-      onChange={onChange}
-      options={options}
-      defaultValue={checkboxes}
-    />
-  );
+export default class CheckboxMenu extends React.Component {
+  static propTypes = {
+    checkboxes: PropTypes.array,
+    label: PropTypes.string,
+    onChange: PropTypes.func,
+    options: PropTypes.array,
+  };
 
-  return (
-    <Dropdown overlay={menu}>
-      <Button icon="bars">{label}</Button>
-    </Dropdown>
-  );
+  static defaultProps = {
+    checkboxes: null,
+    label: null,
+    onChange: null,
+    options: null,
+  };
+
+  render() {
+    const { checkboxes, label, onChange, options } = this.props;
+
+    const menu = (
+      <Checkbox.Group
+        onChange={onChange}
+        options={options}
+        defaultValue={checkboxes}
+      />
+    );
+
+    return (
+      <Dropdown overlay={menu}>
+        <Button icon="bars">{label}</Button>
+      </Dropdown>
+    );
+  }
 }
-
-CheckboxMenu.propTypes = {
-  checkboxes: PropTypes.array,
-  label: PropTypes.string,
-  onChange: PropTypes.func,
-  options: PropTypes.array,
-};
-
-CheckboxMenu.defaultProps = {
-  checkboxes: null,
-  label: null,
-  onChange: null,
-  options: null,
-};
