+++ recipe-server/client/control_new/components/extensions/ExtensionForm.js
@@ -4,8 +4,9 @@ import { is, Map } from 'immutable';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import { createForm, FormItem } from 'control_new/components/common/forms';
-import FormActions from 'control_new/components/common/FormActions';
+import FormActions from 'control_new/components/forms/FormActions';
+import FormItem from 'control_new/components/forms/FormItem';
+import { createForm } from 'control_new/utils/forms';
 
 
 /**
