--- recipe-server/client/control_new/components/pages/Dummy.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function Dummy() {
-  return (
-    <div>Hello world!</div>
-  );
-}
