+++ recipe-server/client/control_new/components/recipes/RecipeForm.js
@@ -5,13 +5,15 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import { createForm, FormItem } from 'control_new/components/common/forms';
-import FormActions from 'control_new/components/common/FormActions';
+import FormItem from 'control_new/components/forms/FormItem';
+import FormActions from 'control_new/components/forms/FormActions';
 import ConsoleLogFields from 'control_new/components/recipes/ConsoleLogFields';
 import PreferenceExperimentFields from 'control_new/components/recipes/PreferenceExperimentFields';
 import ShowHeartbeatFields from 'control_new/components/recipes/ShowHeartbeatFields';
 import { getAction, getAllActions } from 'control_new/state/actions/selectors';
 import { areAnyRequestsInProgress } from 'control_new/state/requests/selectors';
+import { createForm } from 'control_new/utils/forms';
+
 
 /**
  * Form for editing recipes.
